LegendT2dm-package {LegendT2dm}	R Documentation

LegendT2dm: Large-Scale Evidence Generation and Evaluation in a Network of Databases

Description

Large-Scale Evidence Generation and Evaluation in a Network of Databases for Type 2 Diabetes (LEGEND-T2DM) studies the comparative effectiveness and safety of second-line T2DM agents. This systematic, large-scale data experiment follows Open Science principles and aims to deliver reliable and actionable healthcare information for T2DM stakeholders, including patients, their care-givers and physicians.

[Package LegendT2dm version 2.1.1 Index]