addDatabaseIdToTables {LegendT2dm}	R Documentation

Add databaseId column to tables in results *.zip files

Description

Adds databaseId column to tables that are missing such a column in results *.zip files

Usage

addDatabaseIdToTables(
  tableName,
  originalZipFileName,
  newZipFileName,
  databaseId,
  tempFolder = tempdir()
)

Arguments

originalZipFileName
Name of single or vector of multiple zip files.

databaseId
Name or single or vector of multiple databaseId to add

tempFolder
Directory which to use to unpack files

newZipFilaName
Name of single or vector of multiple zip files

[Package LegendT2dm version 2.0.0 Index]