createDataModelSqlFile {LegendT2dm}	R Documentation

Create a SQL file to construct data model tables on a database server.

Description

Create a SQL file to construct data model tables on a database server.

Usage

createDataModelSqlFile(specifications, fileName, tab = "    ")

Arguments

specifications
Specifications data table

fileName
Output name for SQL file

tab
Tab characters to use

Details

Only PostgreSQL servers are supported.

[Package LegendT2dm version 1.1.4 Index]