doMetaAnalysis {LegendT2dm}	R Documentation

Perform random-effects meta-analysis on CES results on data server

Description

Perform random-effects meta-analysis on CES results on data server

Usage

doMetaAnalysis(
  connectionDetails,
  indicationId = "class",
  resultsDatabaseSchema,
  maExportFolder,
  maName = "Meta-analysis",
  diagnosticsFilter = NULL,
  maxCores,
  cacheFileName = NULL
)

Arguments

connectionDetails
DatabaseConnector details to public LEGEND-T2DM db server

resultsDatabaseSchema
Schema on the postgres server where the tables have been created.

maExportFolder
A local folder where the meta-anlysis results will be written.

diagnosticsFilter
Table of which target-comparator-outcome-analysis-database tuples pass diagnostics; can be NULL

maxCores
How many parallel cores should be used? If more cores are made available this can speed up the analyses.

cacheFileName
Use (if exists) or cache database results in file

Details

This function executes a meta-analysis across data source results.

[Package LegendT2dm version 2.0.0 Index]