filterByExposureCohortsSize {LegendT2dm}	R Documentation

Filter exposure pairs by size of the cohorts

Description

Filter exposure pairs by size of the cohorts

Usage

filterByExposureCohortsSize(outputFolder, indicationId, minCohortSize)

Arguments

outputFolder
Name of local folder to place results; make sure to use forward slashes (/)

indicationId
A string denoting the indicationId.

minCohortSize
Minimum number of people that have to be in each cohort to keep a pair of cohorts.

[Package LegendT2dm version 2.1.2 Index]