generateAllCohortMethodDataObjects {LegendT2dm}	R Documentation

Construct all cohortMethodData object

Description

Construct all cohortMethodData object

Usage

generateAllCohortMethodDataObjects(
  outputFolder,
  indicationId = "legendt2dm",
  useSample = FALSE,
  restrictToOt1 = FALSE,
  maxCores = 4
)

Arguments

outputFolder
Name of local folder to place results; make sure to use forward slashes (/)

indicationId
A string denoting the indicationId.

useSample
Use the sampled cohort table instead of the main cohort table (for PS model feasibility).

restrictToOt1
Limit CohortMethod data objects to only comparisons that use On-Treatment-1?

maxCores
How many parallel cores should be used? If more cores are made available this can speed up the analyses.

Details

This function constructs all cohortMethodData objects using the data fetched earlier using the fetchAllDataFromServer function.

[Package LegendT2dm version 2.0.0 Index]