prepareStagedExecution {LegendT2dm}	R Documentation

Utility function to split up CES execution, with only a subset of TC pairs in each run

Description

Utility function to split up CES execution, with only a subset of TC pairs in each run

Usage

prepareStagedExecution(
  originalOutputFolder,
  outputFolderHeader = originalOutputFolder,
  indicationId = "drug",
  stages = 10
)

Arguments

originalOutputFolder
Original output folder used when creating cohorts

outputFolderHeader
The staged execution output folders will start with this header name; default is the same as 'originalOutputFolder'

indicationId
Which sub-study to run?

stages
Number of stages for execution?

Details

This function prepares staged CES execution by copying/slicing necessary files. But note, this does NOT run the study; this is only a utility function that allows staged execution.

[Package LegendT2dm version 3.0.0 Index]