removeResultsFromDatabase {LegendT2dm}	R Documentation

Remove results from the database server.

Description

Requires the results data model tables have been created using the createResultsDataModel function.

Set the POSTGRES_PATH environmental variable to the path to the folder containing the psql executable to enable bulk upload (recommended).

Usage

removeResultsFromDatabase(
  connectionDetails,
  schema,
  databaseId,
  specifications
)

Arguments

connectionDetails
Object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package.

schema
Schema on the postgres server where the tables have been created.

databaseId
Data sources to remove

specifications
Specification of results tables

[Package LegendT2dm version 2.1.0 Index]