unnumberAdditionalCriteria {LegendT2dm}	R Documentation

Find the subsections corresponding to additional criteria and unnumber them.

Description

Find the subsections corresponding to additional criteria and unnumber them.

Usage

unnumberAdditionalCriteria(markdown)

[Package LegendT2dm version 2.1.0 Index]